+++ RegularAverageStandardPong/Assets/Material.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3341e0323ea20d6479f92ea832ccb1fe
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
