+++ RegularAverageStandardPong/Assets/Scripts/QuitScript.cs
