+++ RegularAverageStandardPong/Assets/Scripts/QuitScript.cs.meta
