+++ RegularAverageStandardPong/Assets/Scripts/StartScript.cs
