+++ RegularAverageStandardPong/Assets/Scripts/StartScript.cs.meta
